POF commit
additions to the POF project. 2 games (Windows Phone 8.1) to change and
add to the POF project. One requires putting same numbers displayed on
the screen, the other one is to find and tap on blue rectangles. Both
games are challenges and if solved right will allow user to snooze
alarm.