commit to UserControlSoundSelect
loading with already pre-selected sound. PopUp displays sound names and
plays when "play" button clicked. Gives a choice of selecting a sound
from phone's KnownFolders(Music library, downloads, pictures...) and add
it to collection. If alarm sound was picked up from one of those folders
and than saved it's title will be displayed in the popup otherwise
removed if viewmodel can't locate it in original location.